🚚 [projects] Move module `{templates.adapters.cookiecutter.config => projects}.cookiecutter`
