+++ src/cutty/projects/cookiecutter.py
